alterações no código da home; html, js e css
