+++ app/views/pages/home.html.erb
@@ -1,78 +1,63 @@
-<!DOCTYPE html>
-<html lang="pt-br">
 
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="XÚA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, inicial-scale=1.0">
-  <title>Sidebar Menu using HTML and CSS | Codehal</title>
-  <link rel="stylesheet" href="style.css">
-  <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-</head>
+<div class="carousel" data-controller="carousel">
+  <div class="img-box">
+    <div class="img-list">
+      <div class="img-slider" data-carousel-target="imgSlider">
 
-<body>
-  <div class="carousel">
-    <div class="img-box">
-      <div class="img-list">
-        <div class="img-slider">
-
-          <div class="img-item fruit", style="--i:0;">
-            <%= image_tag(src = "bg/cactus.png") %>
-          </div>
-          <div class="img-item fruit", style="--i:1;">
-            <%= image_tag(src = "bg/cactus.png") %>
-          </div>
-          <div class="img-item fruit", style="--i:2;">
-            <%= image_tag(src = "bg/cactus.png") %>
-          </div>
-          <div class="img-item fruit", style="--i:3;">
-            <%= image_tag(src = "bg/cactus.png") %>
-          </div>
-
-          <div class="img-item", style="--i:0;">
-            <%= image_tag(src = "logo.png") %>
-          </div>
-          <div class="img-item", style="--i:1;">
-            <%= image_tag(src = "BallAndCactus2.png") %>
-          </div>
-          <div class="img-item", style="--i:2;">
-            <%= image_tag(src = "logo.png") %>
-          </div>
-          <div class="img-item", style="--i:3;">
-            <%= image_tag(src = "BallAndCactus2.png") %>
-          </div>
-
+        <div class="img-item fruit", style="--i:0;">
+          <%= image_tag "bg/cactus.png" %>
         </div>
+        <div class="img-item fruit", style="--i:1;">
+          <%= image_tag "bg/cactus.png" %>
+        </div>
+        <div class="img-item fruit", style="--i:2;">
+          <%= image_tag "bg/cactus.png" %>
+        </div>
+        <div class="img-item fruit", style="--i:3;">
+          <%= image_tag "bg/cactus.png" %>
+        </div>
+
+        <div class="img-item", style="--i:0;">
+          <%= image_tag "logo.png" %>
+        </div>
+        <div class="img-item", style="--i:1;">
+          <%= image_tag "BallAndCactus2.png" %>
+        </div>
+        <div class="img-item", style="--i:2;">
+          <%= image_tag "logo.png" %>
+        </div>
+        <div class="img-item", style="--i:3;">
+          <%= image_tag "BallAndCactus2.png" %>
+        </div>
+
       </div>
     </div>
+  </div>
 
-    <div class="info-box">
-      <div class="info-slider">
-        <div class="info-item">
-        <h4>Teste do Teste</h4>
-        <p>Lorem ipsum dolor sit amet consectetur. Reiciendis consequuntur voluptas aliquid ex cumque placeat architecto neque saepe eligendi.</p>
-        </div>
+  <div class="info-box">
+    <div class="info-slider">
+      <div class="info-item">
+      <h4>Teste do Teste</h4>
+      <p>Lorem ipsum dolor sit amet consectetur. Reiciendis consequuntur voluptas aliquid ex cumque placeat architecto neque saepe eligendi.</p>
       </div>
     </div>
-
-    <div class="navigation">
-      <span class="prev-btn">
-        <i class='bx bx-chevron-left'></i>
-      </span>
-
-      <span class="next-btn">
-        <i class='bx bx-chevron-right' ></i>
-      </span>
-    </div>
-
   </div>
 
-  <script src="script.js"></script>
-</body>
+  <div class="navigation">
+    <span class="prev-btn" data-action="carousel#prevBtn">
+      <i class='bx bx-chevron-left'></i>
+    </span>
 
-  <div class="button-group">
-    <%= link_to "ACESSAR", new_user_session_path, class: "btn btn-primary" %>
-    <%= link_to "CADASTRAR-SE", new_user_registration_path, class: "btn btn-outline-primary" %>
+    <span class="next-btn"data-action="carousel#nextBtn">
+      <i class='bx bx-chevron-right' ></i>
+    </span>
   </div>
+
 </div>
-</html>
+
+
+
+<div class="button-group">
+  <%= link_to "ACESSAR", new_user_session_path, class: "btn btn-primary" %>
+  <%= link_to "CADASTRAR-SE", new_user_registration_path, class: "btn btn-outline-primary" %>
+</div>
